renamed AddClusters to AddCluster
